+++ src/app/pages/submission/submission-list/ag-grid/status-cell.component.css
@@ -1,0 +1,3 @@
+.status-badge {
+  background-color: #3c739f;
+}
