+++ src/app/pages/submission/submission-list/ag-grid/text-filter.component.ts
@@ -20,7 +20,6 @@ export class TextFilterComponent implements AgFilterComponent {
   private prev: string = '';
   private valueGetter?: (rowNode: RowNode) => any;
 
-
   afterGuiAttached(): void {
     this.inputEl.nativeElement.focus();
   }
