+++ src/app/pages/submission/submission-list/submission-list.module.ts
@@ -1,13 +1,15 @@
 import { NgModule } from '@angular/core';
 import { AgGridModule } from 'ag-grid-angular';
+import { ButtonsModule } from 'ngx-bootstrap';
+import { ActionButtonsCellComponent } from './ag-grid/action-buttons-cell.component';
+import { DateCellComponent } from './ag-grid/date-cell.component';
+import { DateFilterComponent } from './ag-grid/date-filter.component';
 import { SharedModule } from 'app/shared/shared.module';
+import { StatusCellComponent } from './ag-grid/status-cell.component';
+import { SubmListComponent } from './subm-list.component';
 import { SubmissionOtherModule } from '../submission-other/submission-other.module';
 import { SubmissionSharedModule } from '../submission-shared/submission-shared.module';
-import { DateFilterComponent } from './ag-grid/date-filter.component';
 import { TextFilterComponent } from './ag-grid/text-filter.component';
-
-import { DateCellComponent, ActionButtonsCellComponent, SubmListComponent } from './subm-list.component';
-import { ButtonsModule } from 'ngx-bootstrap';
 
 @NgModule({
   imports: [
@@ -16,7 +18,8 @@ import { ButtonsModule } from 'ngx-bootstrap';
       ActionButtonsCellComponent,
       DateCellComponent,
       TextFilterComponent,
-      DateFilterComponent
+      DateFilterComponent,
+      StatusCellComponent
     ]),
     SubmissionSharedModule,
     SubmissionOtherModule,
@@ -25,6 +28,7 @@ import { ButtonsModule } from 'ngx-bootstrap';
   providers: [],
   declarations: [
     SubmListComponent,
+    StatusCellComponent,
     TextFilterComponent,
     DateFilterComponent,
     ActionButtonsCellComponent,
